moved signal to separate module
